Added call to setupThermistors() which is now needed as we can be using different thermistors at teh same time. ( or even non-thermistors  with thermistors, etc).
Also been hacking on the validate_hardware() function a bit.  It's still very hackish and needs more love, but it's output is now a tad more compatible with the recent protocol changes ( ie firmware output that should be essentially ignored by the host starts with double slashes as comment indicators).